requirements.txt
python-telegram-bot>=21.0
httpx>=0.27.0
matrix-nio[e2e]>=0.27.0
